requirements.txt
amqp==2.5.2
asgiref==3.2.3
Babel==2.7.0
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
crayons==0.3.0
Django==2.2.7
django-celery-results==1.1.2
django-elasticsearch-dsl==7.1.0
django-filter==2.2.0
djangorestframework==3.11.0
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
flower==0.9.3
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
Markdown==3.1.1
more-itertools==8.0.2
numpy==1.17.4
pandas==0.25.3
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python3-logstash==0.4.80
pytz==2019.3
redis==3.3.11
requests==2.22.0
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
tornado==5.1.1
urllib3==1.25.7
vine==1.3.0
webdriver-manager==2.3.0
zipp==0.6.0
